requirements.txt
alembic==0.8.2.dev0
asgiref==2.3.2
async-timeout==3.0.1
attrs==18.2.0
autobahn==18.12.1
Automat==0.7.0
Babel==1.3
blinker==1.3
certifi==2018.11.29
channels==2.1.6
chardet==3.0.4
constantly==15.1.0
cryptography==1.2.3
daphne==2.2.4
defer==1.0.6
dj-database-url==0.5.0
Django==2.1.4
django-cors-headers==2.4.0
django-earthdistance==1
django-heroku==0.3.1
django-model-utils==3.1.2
django-rest-framework==0.1.0
djangorestframework==3.9.0
drf-writable-nested==0.5.1
ecdsa==0.13
Flask==0.10.1
Flask-BabelEx==0.9.3
Flask-Gravatar==0.4.2
Flask-HTMLmin==1.3.2
Flask-Login==0.4.0
Flask-Mail==0.9.1
Flask-Migrate==1.7.0
Flask-Paranoid==0.2.0
Flask-Principal==0.4.0
Flask-Script==0.6.7
Flask-Security==1.7.5
Flask-SQLAlchemy==1.0
Flask-WTF==0.14.2
htmlmin==0.1.12
hyperlink==18.0.0
idna==2.8
incremental==17.5.0
itsdangerous==0.24
Jinja2==2.8
Mako==1.0.3
MarkupSafe==0.23
paramiko==1.16.0
passlib==1.6.5
psutil==3.4.2
psycopg2==2.7.5
psycopg2-binary==2.7.6.1
pyasn1==0.1.9
pycrypto==2.6.1
pycurl==7.43.0
pygobject==3.20.0
PyHamcrest==1.9.0
pyinotify==0.9.6
pyOpenSSL==0.15.1
python-apt==1.1.0b1+ubuntu0.16.4.2
python-dateutil==2.4.2
python-debian==0.1.27
pytz==2014.10
requests==2.21.0
simplejson==3.8.1
six==1.12.0
SQLAlchemy==1.0.11
sqlparse==0.2.4
sshtunnel==0.1.4
Twisted==18.9.0
txaio==18.8.1
typecatcher==0.2
ufw==0.35
unattended-upgrades==0.1
urllib3==1.24.1
Werkzeug==0.10.4
whitenoise==4.1.2
WTForms==2.1
zope.interface==4.6.0
